requirements.txt
langchain-cohere
langchain-huggingface
langchain-ollama
langchain-openai
langchain-community
langchain-core
langchain-text-splitters
langgraph
yt-dlp
scholarly
youtube-transcript-api
pydantic
Flask
croniter
black
